// src/app/shared/model/page.model.ts
export interface CompanyInfo {
  sNo?: number;
  isSelected: boolean;
  CompanyName: string;
  BillCycle: string;
  PaymentMethod: string;
  Email: string;
  AccountURL: string;
  Plan: string;
  CreatedDate: string;
  ExpiringDate: string;
  Image: string;
  Status: string;
  Amount: string;
  DomainStatus: string;
  InvoiceID: string;
}